+++ src/main/java/com/buuz135/simpleclaims/Main.java
@@ -7,6 +7,7 @@ import com.buuz135.simpleclaims.config.SimpleClaimsConfig;
 import com.buuz135.simpleclaims.map.SimpleClaimsChunkWorldMap;
 import com.buuz135.simpleclaims.map.SimpleClaimsWorldMapProvider;
 import com.buuz135.simpleclaims.systems.events.*;
+import com.buuz135.simpleclaims.systems.tick.EntryTickingSystem;
 import com.buuz135.simpleclaims.systems.tick.TitleTickingSystem;
 
 import com.buuz135.simpleclaims.systems.tick.WorldMapUpdateTickingSystem;
@@ -52,6 +53,8 @@ public class Main extends JavaPlugin {
         this.getEntityStoreRegistry().registerSystem(new InteractEventSystem());
         this.getEntityStoreRegistry().registerSystem(new PickupInteractEventSystem());
         this.getEntityStoreRegistry().registerSystem(new TitleTickingSystem(CONFIG.get().getTitleTopClaimTitleText()));
+        if (CONFIG.get().isEnableAlloyEntryTesting())
+            this.getEntityStoreRegistry().registerSystem(new EntryTickingSystem());
         this.getEntityStoreRegistry().registerSystem(new CustomDamageEventSystem());
         this.getChunkStoreRegistry().registerSystem(new WorldMapUpdateTickingSystem());
         this.getCommandRegistry().registerCommand(new SimpleClaimProtectCommand());